remove IComms from public headers
