+++ src/boards/LimeSDR/LimeSDR.cpp
@@ -12,7 +12,7 @@
 #include "lms7002m/LMS7002M_validation.h"
 #include "protocols/LMS64CProtocol.h"
 #include "limesuiteng/DeviceNode.h"
-#include "limesuiteng/IComms.h"
+#include "comms/IComms.h"
 #include "FX3/FX3.h"
 #include "ISerialPort.h"
