+++ src/boards/LimeSDR_Mini/FPGA_Mini.cpp
@@ -1,6 +1,6 @@
 #include "FPGA_Mini.h"
 #include "limesuiteng/Logger.h"
-#include "limesuiteng/IComms.h"
+#include "comms/IComms.h"
 #include <ciso646>
 #include <vector>
 #include <map>
