+++ src/boards_wxgui/pnlX3.cpp
@@ -10,7 +10,7 @@
 #include "lms7suiteEvents.h"
 #include "limesuiteng/SDRDevice.h"
 #include "limesuiteng/SDRDescriptor.h"
-#include "limesuiteng/IComms.h"
+#include "comms/IComms.h"
 
 #include <ciso646>
