+++ src/comms/PCIe/PCIE_CSR_Pipe.h
@@ -1,7 +1,7 @@
 #ifndef PCIE_CSR_PIPE_H
 #define PCIE_CSR_PIPE_H
 
-#include "limesuiteng/IComms.h"
+#include "comms/IComms.h"
 #include "ISerialPort.h"
 #include "LitePCIe.h"
