+++ src/comms/USB/LMS64C_FPGA_Over_USB.h
@@ -1,7 +1,7 @@
 #ifndef LIME_LMS64C_FPGA_OVER_USB_H
 #define LIME_LMS64C_FPGA_OVER_USB_H
 
-#include "limesuiteng/IComms.h"
+#include "comms/IComms.h"
 #include "USB_CSR_Pipe.h"
 #include <memory>
