+++ src/lms7002m/LMS7002M.cpp
@@ -29,7 +29,7 @@
 
 #include "lms_gfir.h"
 #include "limesuiteng/commonTypes.h"
-#include "limesuiteng/IComms.h"
+#include "comms/IComms.h"
 #include "LMS7002M_RegistersMap.h"
 #include "limesuiteng/Logger.h"
 #include "mcu_programs.h"
