+++ src/protocols/LMS64CProtocol.cpp
@@ -6,7 +6,7 @@
 
 #include "limesuiteng/Logger.h"
 #include "limesuiteng/SDRDescriptor.h"
-#include "limesuiteng/IComms.h"
+#include "comms/IComms.h"
 #include "ISerialPort.h"
 #include "LMS64CProtocol.h"
 #include <chrono>
